[cleanup] Add HelloWorldCustomJsonMapper in Java & Kotlin
